requirements.txt
pyyaml
numpy>=1.14.2
pandas>=0.24.0
scipy>=1.0.0
scikit-learn>=0.17.1
rasterio>=1.0.7
rasterstats>=0.13.0
shapely>=1.6.4
geopandas>=0.4.0
